feat(stytch): made login work

+++ frontEnd/src/pages/Authentication/Logout.tsx
@@ -1,20 +1,31 @@
-import { useAuth0 } from '@auth0/auth0-react';
-import { reset } from 'instrumentation/analytics';
 import React from 'react';
+import { useNavigate } from 'react-router-dom';
 import { auth } from 'utils/spaUrls';
+import Spinner from 'components/Common/Spinner';
+import { Logger } from 'utils/logger';
+import { ErrorCodes } from 'utils/errorUtils';
+import { useStytch } from '@stytch/react';
 
 
 export default function Logout() {
-  const { logout } = useAuth0();
-  const clientId = process.env.REACT_APP_AUTH0_CLIENT_ID || '';
-  React.useEffect(() => {
-    localStorage.clear();
-    reset();
-    logout({
-      clientId,
-      logoutParams: { returnTo: `${window.location.origin}${auth.login}` },
-    });
-  }, [clientId, logout]);
+  const stytchClient = useStytch();
+  const navigate = useNavigate();
+  const hasLoggedOut = React.useRef(false);
 
-  return null;
+  React.useEffect(() => {
+    if (!hasLoggedOut.current) {
+      localStorage.clear();
+      stytchClient.session.revoke()
+        .then((resp) => resp.status_code === 200 && navigate(auth.login))
+        .catch((err) => {
+          if (err instanceof Error) {
+            Logger.error(ErrorCodes.auth.stytchGetToken, err.message);
+            navigate(auth.login);
+          }
+        });
+      hasLoggedOut.current = true;
+    }
+  }, [navigate, stytchClient.session]);
+
+  return <Spinner />;
 }
